# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Cooperative Services API",
    "description" : "Runnable prototype. Development OTP and simulated payments are explicitly gated. See docs/api-contract.md for response examples and integration rules.",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/me/logout" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "logout",
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/auth/verify" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "verify_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Verify"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "refresh",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Refresh"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/challenges" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "challenge",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Challenge"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me" : {
      "get" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "me",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "profile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Profile"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/workers/me" : {
      "get" : {
        "tags" : [ "worker-controller" ],
        "operationId" : "profile_1",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/events" : {
      "get" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "events",
        "parameters" : [ {
          "name" : "after",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "default" : 0,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/v1/catalog/subservices" : {
      "get" : {
        "tags" : [ "catalog-controller" ],
        "operationId" : "services",
        "parameters" : [ {
          "name" : "categoryId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "additionalProperties" : { }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Verify" : {
        "type" : "object",
        "properties" : {
          "challengeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "code" : {
            "type" : "string",
            "pattern" : "[0-9]{6}"
          },
          "role" : {
            "type" : "string",
            "pattern" : "CUSTOMER|WORKER"
          },
          "name" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          }
        },
        "required" : [ "challengeId", "code", "name", "role" ]
      },
      "Refresh" : {
        "type" : "object",
        "properties" : {
          "refreshToken" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          }
        },
        "required" : [ "refreshToken" ]
      },
      "Challenge" : {
        "type" : "object",
        "properties" : {
          "phone" : {
            "type" : "string",
            "pattern" : "\\+[1-9][0-9]{7,14}"
          }
        },
        "required" : [ "phone" ]
      },
...
      "Profile" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "preferredLang" : {
            "type" : "string",
            "pattern" : "en|hi|ta"
          }
        },
        "required" : [ "name", "preferredLang" ]
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}